Updating create_patch to start_patch
start_patch takes first class citizens (classes, modules) as opposed to
strings and intelligently builds the patch string needed for the patch
operation. This eliminates the need for the user to know how to
construct the patch string, a process that is often non-obvious and
frought with easy to make mistakes (bad path, typos, etc).

Users can skip the first two arguments to start_patch in favor of the
"target" kwarg, with which a patch string can be passed in as before.